backend/requirements.txt
flask
flask-cors
scapy
netifaces
python-nmap
